+++ src/agoradatatools/etl/transform/__init__.py
@@ -24,9 +24,6 @@ from agoradatatools.etl.transform.disease_correlation import (
 from agoradatatools.etl.transform.model_overview import transform_model_overview
 from agoradatatools.etl.transform.rna_de_aggregate import transform_rna_de_aggregate
 from agoradatatools.etl.transform.rna_de_individual import transform_rna_de_individual
-from agoradatatools.etl.transform.rna_de_individual_chunked import (
-    transform_rna_de_individual_chunked,
-)
 
 
 __all__ = [
@@ -46,5 +43,4 @@ __all__ = [
     "transform_model_overview",
     "transform_rna_de_aggregate",
     "transform_rna_de_individual",
-    "transform_rna_de_individual_chunked",
 ]
